refactor: delete userService to replace functions into user service
- Renamed POST imports for consistency (e.g., LoginPost -> Login).
- Added `getUserSelfData` method to fetch user data.
- Added `logout` method to handle user logout.
- Removed deprecated `userService/UserService.ts` file.
- Updated references to use the new `UserService` structure.
