+++ src/lib/services/user/UserService.ts
@@ -1,11 +1,13 @@
-import { LoginRequestBodySchema, POST as LoginPost } from '@/app/api/auth/login/route'
+import { LoginRequestBodySchema, POST as Login } from '@/app/api/auth/login/route'
 import { buildNextRequestURL } from '@/utils/buildNextRequestURL'
 import { buildNextRequest } from '@/utils/buildNextRequest'
 import { type StatusCodes } from 'http-status-codes'
 import { type typeToFlattenedError } from 'zod'
-import { RegisterRequestBodySchema, POST as RegisterPost } from '@/app/api/auth/register/route'
+import { RegisterRequestBodySchema, POST as Register } from '@/app/api/auth/register/route'
+import { GET as Logout } from '@/app/api/auth/logout/route'
 import { UserRoleWithoutAdmin } from '@/types/User'
-
+import { UserCombinedInfo } from '@/types/Collections'
+import { GET as GetUserSelfData } from '@/app/api/users/me/route'
 const UserService = {
   /**
    * Logs in a user with the provided email and password.
@@ -26,7 +28,7 @@ const UserService = {
     details?: typeToFlattenedError<typeof LoginRequestBodySchema>
   }> {
     const url = buildNextRequestURL('/api/auth/login', options)
-    const response = await LoginPost(await buildNextRequest(url, { method: 'POST', body: options }))
+    const response = await Login(await buildNextRequest(url, { method: 'POST', body: options }))
     const { message, redirect, error, details } = await response.json()
 
     return { message, redirect, status: response.status, error, details }
@@ -55,12 +57,35 @@ const UserService = {
     details?: typeToFlattenedError<typeof RegisterRequestBodySchema>
   }> {
     const url = buildNextRequestURL('/api/auth/register', options)
-    const response = await RegisterPost(
-      await buildNextRequest(url, { method: 'POST', body: options }),
-    )
+    const response = await Register(await buildNextRequest(url, { method: 'POST', body: options }))
     const { message, error, details } = await response.json()
 
     return { message, status: response.status, error, details }
+  },
+
+  /**
+   * Method to get a users own data
+   *
+   * @returns A promise resolving to an object containing:
+   *   - `user` {@link UserCombinedInfo}: The user data, combined with any client additional info.
+   *   - `status` {@link StatusCodes}: The HTTP status code of the response.
+   */
+  getUserSelfData: async function (): Promise<{ user: UserCombinedInfo; status: StatusCodes }> {
+    'use server'
+    const url = buildNextRequestURL('/api/users/me', {})
+    const response = await GetUserSelfData(await buildNextRequest(url, { method: 'GET' }))
+    const { data } = await response.json()
+
+    return { user: data, status: response.status }
+  },
+
+  /**
+   * Logs a user out by clearing their cookies.
+   */
+  logout: async function (): Promise<void> {
+    'use server'
+    const url = buildNextRequestURL('/api/auth/logout', {})
+    await Logout(await buildNextRequest(url, { method: 'GET' }))
   },
 } as const
